refactor: isolate sparse index dcp top-k
